further minor fixes

+++ app/models/Show.scala
@@ -58,7 +58,7 @@ object Show {
     }
   }
 
-  def findShowById(showId: Int) = {
+  def findShowById(showId: Long) = {
     showsCollection.
       // find all people with name `name`
       find(
